+++ srcs/died.c
@@ -1,11 +1,11 @@
 #include "philo.h"
 
-int	is_dead(t_data *data)
+int64_t	is_dead(t_data *data)
 {
 	return (mtx_do_func(data, &data->mdied, dead_check));
 }
 
-int	dead_check(t_data *data)
+int64_t	dead_check(t_data *data)
 {
 	return (data->died + data->phi->end_of_simulation);
 }
@@ -15,7 +15,7 @@ void	i_have_died(t_data *data)
 	mtx_do_func(data, &data->mdied, set_deadflag);
 }
 
-int	set_deadflag(t_data *data)
+int64_t	set_deadflag(t_data *data)
 {
 	data->died = 1;
 	data->phi->end_of_simulation = 1;
